pyproject.toml
[project]
name = "deep-research"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "openai>=2.14.0",
    "pydantic-settings>=2.12.0",
    "structlog>=25.5.0",
    "tavily>=1.1.0",
    "tenacity>=9.0.0",
]

[dependency-groups]
dev = [
    "basedpyright>=1.36.2",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "ruff>=0.14.10",
]

[tool.basedpyright]
include = ["src"]
reportAny = false
reportExplictAny = false

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = [
  "ANN",
  "ASYNC",
  "B",
  "C4",
  "E",
  "F",
  "FAST",
  "I",
  "PIE",
  "PTH",
  "RUF",
  "S",
  "SIM",
  "T20",
  "UP",
  "W",
]
ignore = [
  "ANN401",
  "COM812",
  "E501",
  "ISC001",
  "B008",
]
[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
  "ANN",
  "S"
]
